@@ -17,26 +17,26 @@
     "@material-ui/icons": "^1.1.0",
     "amazon-cognito-identity-js": "^2.0.13",
     "apollo-client": "^2.3.5",
-    "aws-appsync": "^1.1.4",
-    "aws-appsync-react": "^1.0.19",
+    "aws-appsync": "^1.2.1",
+    "aws-appsync-react": "^1.0.21",
     "graphql": "^0.13.2",
     "graphql-tag": "^2.9.2",
     "moment": "^2.19.2",
     "react": "^16.4.0",
-    "react-apollo": "^2.1.8",
+    "react-apollo": "^2.1.9",
     "react-dom": "^16.4.0",
     "react-router-dom": "^4.2.2",
-    "react-swipeable-views": "^0.12.13",
+    "react-swipeable-views": "^0.12.14",
     "styled-components": "3.3.3",
     "uuid": "^3.3.2"
   },
   "devDependencies": {
-    "@babel/core": "v7.0.0-beta.51",
-    "@babel/node": "7.0.0-beta.51",
-    "@babel/preset-env": "v7.0.0-beta.51",
-    "@babel/preset-react": "v7.0.0-beta.51",
-    "@babel/preset-stage-3": "^7.0.0-beta.51",
-    "@babel/preset-typescript": "v7.0.0-beta.51",
+    "@babel/core": "v7.0.0-beta.53",
+    "@babel/node": "7.0.0-beta.53",
+    "@babel/preset-env": "v7.0.0-beta.53",
+    "@babel/preset-react": "v7.0.0-beta.53",
+    "@babel/preset-stage-3": "^7.0.0-beta.53",
+    "@babel/preset-typescript": "v7.0.0-beta.53",
     "@types/react": "^16.4.6",
     "@types/react-dom": "^16.0.5",
     "@types/react-router-dom": "^4.2.6",
@@ -51,9 +51,9 @@
     "tslint-loader": "^3.6.0",
     "tslint-react": "^3.6.0",
     "typescript": "^2.9.2",
-    "typescript-styled-plugin": "^0.7.0",
-    "webpack": "^4.14.0",
+    "typescript-styled-plugin": "^0.10.0",
+    "webpack": "^4.16.0",
     "webpack-cli": "^3.0.8",
-    "webpack-serve": "^1.0.2"
+    "webpack-serve": "^2.0.2"
   }
 }
