@@ -1,5 +1,5 @@
 import styled from "styled-components";
 
 export default styled.div`
-    margin-top: 7rem;
+    min-height: inherit;
 `;
